redux uploaded

+++ react/reduxcrud/src/Pages/Viewuser.jsx
@@ -1,19 +1,24 @@
 import React,{useEffect} from 'react'
-import {Link, useNavigate} from 'react-router-dom'
-import {useSelector,useDispatch} from 'react-redux'
+import { useDispatch, useSelector } from 'react-redux'
+import { Link, useNavigate } from 'react-router-dom'
+import { deleteUser, loadUsers } from '../Redux/actions';
 
-import { getUser,deleteUser } from '../Redux/actions'
-function View_user() {
-    const {usersarr}=useSelector((state)=>state.usersR)
-    let dispatch=useDispatch();
-    useEffect(()=>{
-        dispatch(getUser());
-    },[]);
 
-    function handelDelete(id) {
-        dispatch(deleteUser(id))
+const Viewuser = () => {
+
+  const navigate= useNavigate();
+  let dispatch= useDispatch();
+  const {users}= useSelector(state=> state.data);
+
+  useEffect(() => {
+    dispatch(loadUsers())
+  }, [])
+  
+  const handledelete= (id)=>{
+    if(window.confirm("Are You Sure Wanted to Delete the User?")){
+      dispatch(deleteUser(id));
     }
-    const navigate = useNavigate()
+  }
   return (
       <div className="container mt-4">
           <h2>View User</h2>
@@ -26,21 +31,23 @@ function View_user() {
                       <th>Username</th>
                       <th>Email</th>
                       <th>Phone</th>
+                      <th>Website</th>
                       <th>Action</th>
                   </tr>
               </thead>
               <tbody>
                 {
-                   usersarr && usersarr.map((item)=>(
+                   users && users.map((item,index)=>(
                     <tr key={item.id}>
-                      <th>{item.id}</th>
+                      <th>{index+1}</th>
                       <th>{item.name}</th>
                       <th>{item.username}</th>
                       <th>{item.email}</th>
                       <th>{item.phone}</th>
+                      <th>{item.website}</th>
                       <th>
-                        <button className='btn btn-danger m-1' onClick={()=>handelDelete(item.id)}>Delete</button>
-                        <button className='btn btn-primary m-1'onClick={()=>{navigate(`/edit_user/${item.id}`)}}>Edit</button>
+                        <button className='btn btn-danger m-1' onClick={()=> handledelete(item.id)} >Delete</button>
+                        <button className='btn btn-primary m-1' onClick={()=>navigate(`/edit_user/${item.id}`)}>Edit</button>
                       </th>
                   </tr>
                     ))
@@ -49,7 +56,7 @@ function View_user() {
           </table>
       </div>
 
-  )
+    )
 }
 
-export default View_user
+export default Viewuser
